Rename data source directory also
